chore: certificate 스텝별로 분리
