+++ frontend/src/components/certificate/index.ts
@@ -1,0 +1,6 @@
+export { Step1 } from './Step1';
+export { Step2 } from './Step2';
+export { Step3 } from './Step3';
+export { Step4 } from './Step4';
+export { Step5 } from './Step5';
+export { Step6 } from './Step6';
